Move protobuf codegen into its own crate warg-protobuf
