+++ crates/protocol/Cargo.toml
@@ -7,6 +7,7 @@ rust-version = { workspace = true }
 
 [dependencies]
 warg-crypto = { workspace = true }
+warg-protobuf = { workspace = true }
 warg-transparency = { workspace = true }
 wasmparser = { workspace = true }
 thiserror = { workspace = true }
@@ -14,7 +15,6 @@ anyhow = { workspace = true }
 serde = { workspace = true }
 prost = { workspace = true }
 prost-types = { workspace = true }
-pbjson = { workspace = true }
 pbjson-types = { workspace = true }
 hex = { workspace = true }
 base64 = { workspace = true }
@@ -25,8 +25,3 @@ indexmap = { workspace = true }
 [dev-dependencies]
 pretty_assertions = { workspace = true }
 serde_json = { workspace = true }
-
-[build-dependencies]
-prost-build = { workspace = true }
-pbjson-build = { workspace = true }
-regex = { workspace = true }
