+++ crates/transparency/src/lib.rs
@@ -17,9 +17,5 @@ pub mod map;
 
 /// Types for converting to and from protobuf
 pub(crate) mod protobuf {
-    #![allow(clippy::all, unused_qualifications)]
-    // Generated by [`prost-build`]
-    include!(concat!(env!("OUT_DIR"), "/warg.transparency.rs"));
-    // Generated by [`pbjson-build`]
-    include!(concat!(env!("OUT_DIR"), "/warg.transparency.serde.rs"));
+    pub use warg_protobuf::transparency::*;
 }
